Got view for tagging directorships working.

+++ mirror/sql/create_companies_table.sql
@@ -1,4 +1,4 @@
 DROP TABLE IF EXISTS companies;
 
-SELECT DISTINCT regexp_replace(director_id, '\..*$', '') as equilar_id, company
+SELECT DISTINCT regexp_replace(director_id, '\..*$', '')::integer as equilar_id, company
     INTO companies FROM director;
